requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-filter==21.1
django-extensions==3.1.5
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
psycopg2==2.9.3
Pillow==9.1.1
PyYAML==6.0
PyJWT==2.4.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0











